Find Employee logic

+++ src/main/java/io/evilkhaoskat/tests/entity/Employee.java
@@ -1,18 +1,19 @@
 package io.evilkhaoskat.tests.entity;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Entity
 @Table
 public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private BigInteger id;
     private String name;
     private double salary;
     private String jobTitle;
 
-    public Employee(int id, String name, double salary, String jobTitle) {
+    public Employee(BigInteger id, String name, double salary, String jobTitle) {
         this.id = id;
         this.name = name;
         this.salary = salary;
@@ -23,11 +24,11 @@ public class Employee {
         super();
     }
 
-    public int getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
